backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.12.0